ai-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
